% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolor_sampler.R
\name{coolor_sampler}
\alias{coolor_sampler}
\title{Generate a color palette from https://coolors.co/}
\usage{
coolor_sampler(ncolor = 5)
}
\arguments{
\item{ncolor}{Input number of colors in palette (2-9)}
}
\value{
Hex color codes of palette.
}
\description{
This function grabs a randomly generated color palette
of 2 to 9 colors from https://coolors.co/.
}
\examples{
## Generate a palette of 6 colors from https://coolors.co/
my_colors <- coolor_sampler(ncolor=6)

## View sampled colors
scales::show_col(my_colors, ncol = 1)

## Easily add the palette to your next ggplot!
background_col <- my_colors[1]
points_col <- my_colors[2:length(my_colors)]

data("mtcars")
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl))) +
  scale_colour_manual(values = rev(points_col)) +
  theme(plot.background = element_rect(background_col),
        panel.background = element_rect(background_col),
        legend.background = element_rect(background_col),
        text = element_text(color="white"))
}
\keyword{coolors}
